Migrated TOUCH to a separate crate.
