+++ crates/shared/src/constants/general.rs
@@ -1,11 +1,5 @@
 pub const AUTHOR_NAME: &str = "Breno RdV";
 
-pub const NSCV_APP_NAME: &str = "nscv";
-pub const NSCV_VERSION: &str = "1.0.0";
-
-pub const TOUCH_APP_NAME: &str = "touch";
-pub const TOUCH_VERSION: &str = "1.0.0";
-
 pub const TS_APP_NAME: &str = "ts";
 pub const TS_VERSION: &str = "1.1.0";
