+++ crates/tool-touch/Cargo.toml
@@ -1,16 +1,15 @@
 [package]
-name = "csvn"
+name = "touch"
 version = "1.0.0"
 edition = "2021"
 authors = ["Breno RdV"]
-description = "CLI tool that fills missing CSV fields with defaults and outputs a `_normalized` file for consistent, error-free downstream processing."
+description = "Tool for splitting large UTF-8 text or CSV files by line count with header preservation, progress feedback, and graceful shutdown."
 repository = "https://github.com/brenordv/rusted-toolbox"
 
 [dependencies]
 shared = { path = "../shared" }
-string-interner = "0.19.0"
+filetime = "0.2.25"
 anyhow = "1.0.99"
 clap = "4.5.45"
 chrono = "0.4.41"
-csv = "1.3.1"
 tracing = "0.1.41"
