+++ crates/tool-touch/src/touch_app.rs
@@ -1,9 +1,9 @@
-use crate::tools::touch::models::TouchArgs;
 use anyhow::{Context, Result};
 use filetime::{set_file_times, set_symlink_file_times, FileTime};
 use shared::system::get_full_filepath_from_string::get_full_filepath_from_string;
 use std::fs::File;
 use std::path::PathBuf;
+use crate::models::TouchArgs;
 
 /// Creates a file if it doesn't exist based on the no_create flag.
 ///
