--- crates/tools/src/tools/touch/mod.rs
@@ -1,3 +1,0 @@
-pub mod cli_utils;
-pub mod models;
-pub mod touch_app;
